Fix annoying zsh prompt redraw problem on SIGWINCH.
Zsh has to redraw the prompt on SIGWINCH and it does so by clearing to
the start of the current line.  This doesn't play well with multi-line
prompts.  So:

- Keep PS1 a simple one-line prompt that zsh can redraw at will.
- Print multi-line prompt info from precmd().

SIGWINCH-forced redraws that were annoying me toggling fullscreen in
iTerm2 and creating/deleting panes in tmux.

The explanation of zsh's SIGWINCH PS1 redraw (and precmd idea) from
Bart Schaefer: http://www.zsh.org/mla/users/2004/msg00688.html